requirements.txt
faiss==1.5.3
Flask==2.0.3
matplotlib==3.4.2
numpy==1.20.3
opencv_python==4.5.3.56
Pillow==9.0.1
PyMySQL==1.0.2
PyQt5==5.15.6
torch==1.9.0
torchvision==0.10.0
Werkzeug==2.0.3
